+++ OData2Poco.Tests/MetaDataReaderTests.cs
@@ -1,10 +1,15 @@
 ﻿//#define local
+<<<<<<< HEAD
 
+=======
+>>>>>>> develop
 using System;
 using System.IO;
 using System.Net;
 using System.Xml;
 using NUnit.Framework;
+//using System.Net.Http;
+
 
 namespace OData2Poco.Tests
 {
@@ -12,6 +17,7 @@ namespace OData2Poco.Tests
     [TestFixture]
     public partial class MetaDataReaderTests
     {
+<<<<<<< HEAD
         const string UrlV4 = "http://services.odata.org/V4/Northwind/Northwind.svc";
         const string UrlV3 = "http://services.odata.org/V3/Northwind/Northwind.svc";
         private const string LocalUrl = "http://asd-pc/odata2/api/northwind";
@@ -80,16 +86,18 @@ namespace OData2Poco.Tests
             StringAssert.Contains("[Key]", code);
             Assert.AreEqual(metaDataReader.Generate(new PocoSetting()).ClassList.Count, expecteCount);
           }
+=======
+>>>>>>> develop
 
+        //***********************http test*****************************
         [Test]
-        [TestCase(UrlV4, 26)]
-        [TestCase(UrlV3, 26)]
-        [TestCase(@"data\northwindV4.xml", 11)] //filename , countof entities in the model
-        [TestCase(@"data\northwindV3.xml", 11)]
+        [TestCase("http://services.odata.org/V4/OData/OData.svc",11)] //v4
+        [TestCase("http://services.odata.org/V3/OData/OData.svc",11)] //v3
         //expecteCount: count of classes /Entities
-        public void GeneratePocoFromHttporFileGeneratorTest(string url, int expecteCount)
+        public void GeneratePocoFromHttpTest(string url,int expecteCount)
         {
             var metaDataReader = new MetaDataReader(url);
+<<<<<<< HEAD
             var gen = metaDataReader.Generate(new PocoSetting()); //.Generator; //.Execute();
              
             var code = gen.ToString();
@@ -97,6 +105,18 @@ namespace OData2Poco.Tests
             StringAssert.Contains("public class Product", code);
             //Assert.AreEqual(metaDataReader.Generate().GetClassTemplateList().Count, expecteCount);
 
+=======
+            var code = metaDataReader.GeneratePoco();
+            Assert.IsNotEmpty(code);
+            StringAssert.Contains("public class Product", code);
+            StringAssert.Contains("public class FeaturedProduct", code);
+            Assert.AreEqual(metaDataReader.ClassList.Count, expecteCount);
+            //metaDataReader.ClassList.ForEach(m =>
+            //{
+            //    Console.WriteLine(m.Name);
+            //});
+            
+>>>>>>> develop
         }
 
         [Test]
@@ -115,7 +135,8 @@ namespace OData2Poco.Tests
         [Test]
         public void GeneratePocoWithValidAccountTes()
         {
-            var metaDataReader = new MetaDataReader(LocalUrl, "user", "****");
+            string url = "http://localhost/odata2/api/northwind";
+            var metaDataReader = new MetaDataReader(url,"user","password");
             var code = metaDataReader.GeneratePoco();
             Assert.IsNotEmpty(code);
 
@@ -126,13 +147,39 @@ namespace OData2Poco.Tests
         public void GeneratePocoWithInValidAccountTest()
         {
             var msg = "The remote server returned an error: (401) Unauthorized";
-            MetaDataReader metaDataReader = new MetaDataReader(LocalUrl, "user_invalid", "password");
+            string url = "http://alocalhost/odata2/api/northwind";
+            MetaDataReader metaDataReader = new MetaDataReader(url, "user_invalid", "password");
             var ex = Assert.Throws<WebException>(() => metaDataReader.GeneratePoco());
-            Debug.WriteLine(ex.Message);
             StringAssert.Contains(msg, ex.Message);
 
         }
 #endif
+
+        //********************* xml File test **************************************
+        
+        [Test]
+        public void GeneratePocoFromXmlFileV4Test()
+        {
+            string url = @"data\northwindV4.xml";
+            MetaDataReader metaDataReader = new MetaDataReader(url);
+            metaDataReader.LoadMetaData();
+            Assert.IsNotNull(metaDataReader.GeneratePoco());
+            var count = metaDataReader.ClassList.Count;  
+            Console.Write(count);
+            Assert.AreEqual(count, 11);
+        }
+
+
+        [Test]
+        public void GeneratePocoFromXmlFileV3Test()
+        {
+            string url = @"data\northwindV3.xml";
+            MetaDataReader metaDataReader = new MetaDataReader(url);
+          var code=  metaDataReader.GeneratePoco();
+            Assert.IsNotNull(code);
+            var count = metaDataReader.ClassList.Count;  
+            Assert.AreEqual(count, 11);
+        }
 
         [Test]
         public void GeneratePocoFromFileNotExistTest()
@@ -140,11 +187,16 @@ namespace OData2Poco.Tests
             string code = "";
             string url = "file_not_exist";
             MetaDataReader metaDataReader = new MetaDataReader(url);
+<<<<<<< HEAD
            // Assert.Throws<FileNotFoundException>(() => code = metaDataReader.GeneratePoco());
             Assert.Throws<FileNotFoundException>(() => code = metaDataReader.Generate(new PocoSetting()).ToString()); //.GeneratePoco());
+=======
+            Assert.Throws<FileNotFoundException>(() => code = metaDataReader.GeneratePoco());
+           
+>>>>>>> develop
             Assert.IsEmpty(code);
-        }
 
+        }
         [Test]
         public void GeneratePocoFromNotValidXmlFileTest()
         {
@@ -156,6 +208,6 @@ namespace OData2Poco.Tests
 
         }
 
-
+       
     }
 }