+++ OData2PocoLib/Media.cs
@@ -1,9 +1,0 @@
-namespace OData2Poco
-{
-    public enum Media
-    {
-        Http,
-        File,
-        Unknown
-    }
-}